pyproject.toml
[tool.poetry]
name = "hazard-reporting-system-be"
version = "0.1.0"
description = "A Hazard Reporting System"
authors = ["tasfiqulghani", "skfarhad <sk.farhad.eee@gmail.com>"]
license = "NA"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
celery = "^5.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-leaflet = "^0.30.1"
djangorestframework-api-key = "^3.0.0"
drf-yasg = "^1.21.7"
django = ">=4.2.0,<4.3.0"
gevent = "^24.2.1"
django-filter = "^24.3"
firebase-admin = "^6.5.0"
django-cors-headers = "^4.4.0"
django-background-tasks = "^1.2.8"
django-rest-swagger = "^2.2.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
